some comments, VersionTest moved to proper package
